replace unnecessary tsx files
