+++ package.json
@@ -15,7 +15,8 @@
     "riffraff": "node-riffraff-artifact",
     "modules-dev": "rollup --watch --config",
     "modules": "rollup --config",
-    "dev": "yarn modules & ts-node-dev --transpile-only src/server.tsx",
+    "server-dev": "export stage=DEV; ts-node-dev --transpile-only src/server.ts",
+    "dev": "yarn modules & ts-node-dev --transpile-only src/server.ts",
     "test": "jest",
     "lint": "eslint . --ext .js,.jsx,.ts,.tsx",
     "build": "yarn modules && tsc",
